+++ src/components/user-action/Label.tsx
@@ -1,4 +1,5 @@
 import type { LabelHTMLAttributes } from 'react'
+import clsx from 'clsx'
 
 export type LabelType = 'labelSmall' | 'labelMedium' | 'labelBig'
 
@@ -7,7 +8,6 @@ const styleMapping: Record<LabelType, string> = {
   labelMedium: 'textstyle-label-md',
   labelBig: 'textstyle-label-lg',
 }
-
 
 export type LabelProps = {
   /** The text for the label */
@@ -23,11 +23,12 @@ export const Label = ({
                         children,
                         name,
                         labelType = 'labelSmall',
+                        className,
                         ...props
                       }: LabelProps) => {
   return (
-    <label {...props}>
-      {children ? children : (<span className={styleMapping[labelType]}>{name}</span>)}
+    <label {...props} className={clsx(styleMapping[labelType], className)}>
+      {children ? children : name}
     </label>
   )
 }